% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_simulations.R
\name{big_endian_bits_to_int}
\alias{big_endian_bits_to_int}
\title{Big endian bits to int}
\usage{
big_endian_bits_to_int(bits)
}
\arguments{
\item{bits}{Descending bits of the integer, with the 1s bit at the end.}
}
\value{
The integer.
}
\description{
Returns the big-endian integer specified by the given bits.
}
\seealso{
Other Sampling, Simulations and Data collection: 
\code{\link{big_endian_digits_to_int}()},
\code{\link{big_endian_int_to_bits}()},
\code{\link{big_endian_int_to_digits}()},
\code{\link{circuit_like}()},
\code{\link{final_density_matrix}()},
\code{\link{final_wavefunction}()},
\code{\link{flatten_to_ops_or_moments}()},
\code{\link{flatten_to_ops}()},
\code{\link{flatten_with_params}()},
\code{\link{flatten_with_sweep}()},
\code{\link{flatten}()},
\code{\link{hog_score_xeb_fidelity_from_probabilities}()},
\code{\link{measure_density_matrix}()},
\code{\link{measure_state_vector}()},
\code{\link{random_state_or_seed_like}()},
\code{\link{sample_density_matrix}()},
\code{\link{sample_state_vector}()},
\code{\link{sample_sweep}()},
\code{\link{sample}()},
\code{\link{to_resolvers}()},
\code{\link{to_sweep}()},
\code{\link{validate_mixture}()}
}
\concept{Sampling, Simulations and Data collection}
