% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_simulations.R
\name{big_endian_digits_to_int}
\alias{big_endian_digits_to_int}
\title{Big endian digits to int}
\usage{
big_endian_digits_to_int(digits, base)
}
\arguments{
\item{digits}{Digits of the integer, with the least
significant digit at the end.}

\item{base}{The base, or list of per-digit bases, to use
when combining the digits into an integer. When a list of
bases is specified, the last entry in the list is the base
for the last entry of the digits list (i.e. the least
significant digit). That is to say, the bases are also
specified in big endian order.}
}
\value{
The integer.
}
\description{
Returns the big-endian integer specified
by the given digits and base.
}
\section{Raises}{

ValueError: One of the digits is out of range for its base.
The base was specified per-digit (as a list) but the length of
the bases list is different from the number of digits.
}

\seealso{
Other Sampling, Simulations and Data collection: 
\code{\link{big_endian_bits_to_int}()},
\code{\link{big_endian_int_to_bits}()},
\code{\link{big_endian_int_to_digits}()},
\code{\link{circuit_like}()},
\code{\link{final_density_matrix}()},
\code{\link{final_wavefunction}()},
\code{\link{flatten_to_ops_or_moments}()},
\code{\link{flatten_to_ops}()},
\code{\link{flatten_with_params}()},
\code{\link{flatten_with_sweep}()},
\code{\link{flatten}()},
\code{\link{hog_score_xeb_fidelity_from_probabilities}()},
\code{\link{measure_density_matrix}()},
\code{\link{measure_state_vector}()},
\code{\link{random_state_or_seed_like}()},
\code{\link{sample_density_matrix}()},
\code{\link{sample_state_vector}()},
\code{\link{sample_sweep}()},
\code{\link{sample}()},
\code{\link{to_resolvers}()}
}
\concept{Sampling, Simulations and Data collection}
