% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_simulations.R
\name{big_endian_int_to_bits}
\alias{big_endian_int_to_bits}
\title{Big endian int to bits}
\usage{
big_endian_int_to_bits(val, bit_count)
}
\arguments{
\item{val}{The integer to get bits from. This integer is
permitted to be larger than `2**bit_count` (in which case
the high bits of the result are dropped) or to be negative
(in which case the bits come from the 2s complement signed
representation).}

\item{bit_count}{The number of desired bits in the result.}
}
\value{
The bits.
}
\description{
The big-endian bits of an integer.
}
\seealso{
Other Sampling, Simulations and Data collection: 
\code{\link{big_endian_bits_to_int}()},
\code{\link{big_endian_digits_to_int}()},
\code{\link{big_endian_int_to_digits}()},
\code{\link{circuit_like}()},
\code{\link{final_density_matrix}()},
\code{\link{final_wavefunction}()},
\code{\link{flatten_to_ops_or_moments}()},
\code{\link{flatten_to_ops}()},
\code{\link{flatten_with_params}()},
\code{\link{flatten_with_sweep}()},
\code{\link{flatten}()},
\code{\link{hog_score_xeb_fidelity_from_probabilities}()},
\code{\link{measure_density_matrix}()},
\code{\link{measure_state_vector}()},
\code{\link{random_state_or_seed_like}()},
\code{\link{sample_density_matrix}()},
\code{\link{sample_state_vector}()},
\code{\link{sample_sweep}()},
\code{\link{sample}()},
\code{\link{to_resolvers}()},
\code{\link{to_sweep}()},
\code{\link{validate_mixture}()}
}
\concept{Sampling, Simulations and Data collection}
