% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_simulations.R
\name{big_endian_int_to_digits}
\alias{big_endian_int_to_digits}
\title{Big endian int to digits}
\usage{
big_endian_int_to_digits(val, base, digit_count)
}
\arguments{
\item{val}{The integer to get digits from. Must be non-negative
and less than the maximum representable value, given the
specified base(s) and digit count.}

\item{base}{The base, or list of per-digit bases, to separate
val into. When a list of bases is specified, the last entry in
the list is the base for the last entry of the result (i.e. the
least significant digit). That is to say, the bases are also
specified in big endian order.}

\item{digit_count}{The length of the desired result.}
}
\value{
The list of digits.
}
\description{
Separates an integer into big-endian digits.
}
\section{Raises}{

ValueError: Unknown digit count. The `base` was specified as
an integer and a `digit_count` was not provided. Inconsistent
digit count. The `base` was specified as a per-digit list, and
`digit_count` was also provided, but they disagree.
}

\seealso{
Other Sampling, Simulations and Data collection: 
\code{\link{big_endian_bits_to_int}()},
\code{\link{big_endian_digits_to_int}()},
\code{\link{big_endian_int_to_bits}()},
\code{\link{circuit_like}()},
\code{\link{final_density_matrix}()},
\code{\link{final_wavefunction}()},
\code{\link{flatten_to_ops_or_moments}()},
\code{\link{flatten_to_ops}()},
\code{\link{flatten_with_params}()},
\code{\link{flatten_with_sweep}()},
\code{\link{flatten}()},
\code{\link{hog_score_xeb_fidelity_from_probabilities}()},
\code{\link{measure_density_matrix}()},
\code{\link{measure_state_vector}()},
\code{\link{random_state_or_seed_like}()},
\code{\link{sample}()}
}
\concept{Sampling, Simulations and Data collection}
