% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circuits.R
\name{circuit}
\alias{circuit}
\title{Circut}
\usage{
circuit()
}
\value{
operations.
}
\description{
A mutable list of groups of operations to apply to some qubits
}
\seealso{
Other Circuits: 
\code{\link{circuit_dag}()},
\code{\link{gate_operation_}()},
\code{\link{gate_parallel_operation}()},
\code{\link{insert_strategy}()},
\code{\link{moment}()},
\code{\link{qubit_order_or_list}()},
\code{\link{qubit_order}()},
\code{\link{tree_flatten_op}()},
\code{\link{tree_op}()}
}
\concept{Circuits}
