% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_simulations.R
\name{circuit_like}
\alias{circuit_like}
\title{Circuit like}
\usage{
circuit_like(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
A circuits.Circuit or a value that can be
trivially converted into it: a gate, an operation,
and a list or tree of operations.
}
\seealso{
Other Sampling, Simulations and Data collection: 
\code{\link{big_endian_bits_to_int}()},
\code{\link{big_endian_digits_to_int}()},
\code{\link{big_endian_int_to_bits}()},
\code{\link{big_endian_int_to_digits}()},
\code{\link{final_density_matrix}()},
\code{\link{final_wavefunction}()},
\code{\link{flatten_to_ops_or_moments}()},
\code{\link{flatten_to_ops}()},
\code{\link{flatten_with_params}()},
\code{\link{flatten_with_sweep}()},
\code{\link{flatten}()},
\code{\link{hog_score_xeb_fidelity_from_probabilities}()},
\code{\link{measure_density_matrix}()},
\code{\link{measure_state_vector}()},
\code{\link{random_state_or_seed_like}()},
\code{\link{sample_density_matrix}()},
\code{\link{sample_state_vector}()},
\code{\link{sample_sweep}()},
\code{\link{sample}()},
\code{\link{to_resolvers}()}
}
\concept{Sampling, Simulations and Data collection}
