% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circuits.R
\name{cirquit_unique}
\alias{cirquit_unique}
\title{Unique}
\usage{
cirquit_unique(...)
}
\arguments{
\item{...}{parameters to pass.}
}
\value{
None.
}
\description{
A wrapper for a value that doesn't compare equal
to other instances.
}
\details{
For example: 5 == 5 but Unique(5) != Unique(5). Unique is used by
CircuitDag to wrap operations because nodes in a graph
are considered the same node if they compare equal to each other. X(q0)
in one moment of a Circuit and X(q0) in another moment of the Circuit are
wrapped by Unique(X(q0)) so they are distinct nodes in the graph.
}
