% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pauli_Clifford.R
\name{clifford_single_qubit_gate}
\alias{clifford_single_qubit_gate}
\title{Single Qubit Clifford Gate}
\usage{
clifford_single_qubit_gate(...)
}
\arguments{
\item{...}{parameters to pass.}
}
\value{
rotated object.
}
\description{
Any single qubit Clifford rotation.
}
\seealso{
Other Pauli and Clifford groups: 
\code{\link{clifford_state}()},
\code{\link{clifford_tableau}()},
\code{\link{pauli_base_dense_string}()},
\code{\link{pauli_basis}()},
\code{\link{pauli_dense_string}()},
\code{\link{pauli_gate_like}()},
\code{\link{pauli_interaction_gate}()},
\code{\link{pauli_mutable_dense_string}()},
\code{\link{pauli_pow_combination}()},
\code{\link{pauli_string_gate_operation}()},
\code{\link{pauli_string}()},
\code{\link{pauli_sum_like}()},
\code{\link{pauli_sum}()},
\code{\link{pauli_transform}()},
\code{\link{pauli}()}
}
\concept{Pauli and Clifford groups}
