% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pauli_Clifford.R
\name{clifford_state}
\alias{clifford_state}
\title{Clifford State}
\usage{
clifford_state(qubit_map, initial_state = 0)
}
\arguments{
\item{qubit_map}{mapped qubits}

\item{initial_state}{the initial state}
}
\description{
A state of the Clifford simulation.
}
\details{
The state is stored using two complementary representations:
Anderson's tableaux form and Bravyi's CH-form.
The tableaux keeps track of the stabilizer operations, while the
CH-form allows access to the full wavefunction (including phase).
Gates and measurements are applied to each representation in O(n^2) time.
}
\section{The state is stored using two complementary representations}{

Anderson's tableaux form and Bravyi's CH-form. The tableaux keeps track
of the stabilizer operations, while the CH-form allows access to the full
wavefunction (including phase).
}

\seealso{
Other Pauli and Clifford groups: 
\code{\link{clifford_tableau}()},
\code{\link{pauli_base_dense_string}()},
\code{\link{pauli_basis}()},
\code{\link{pauli_dense_string}()},
\code{\link{pauli_gate_like}()},
\code{\link{pauli_interaction_gate}()},
\code{\link{pauli_mutable_dense_string}()},
\code{\link{pauli_pow_combination}()},
\code{\link{pauli}()}
}
\concept{Pauli and Clifford groups}
