% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_simulations.R
\name{final_density_matrix}
\alias{final_density_matrix}
\title{Final density matrix}
\usage{
final_density_matrix(
  program,
  noise,
  param_resolver,
  qubit_order,
  initial_state,
  dtype,
  seed,
  ignore_measurement_results
)
}
\arguments{
\item{program}{The circuit, gate, operation, or tree of
operations to apply to the initial state in order to produce
the result.}

\item{noise}{Noise model to use while running the simulation.}

\item{param_resolver}{Parameters to run with the program.}

\item{qubit_order}{Determines the canonical ordering of the qubits.
This is often used in specifying the initial state, i.e. the
ordering of the computational basis states.}

\item{initial_state}{If an int, the state is set to the
computational basis state corresponding to this state. Otherwise
if this is a np.ndarray it is the full initial state. In this case
it must be the correct size, be normalized (an L2 norm of 1), and
be safely castable to an appropriate dtype for the simulator.}

\item{dtype}{The numpy.dtype used by the simulation.
Typically one of numpy.complex64 or numpy.complex128.}

\item{seed}{The random seed to use for this simulator.}

\item{ignore_measurement_results}{Defaults to True. When TRUE,
the returned density matrix is not conditioned on any measurement
results. For example, this effectively replaces computational basis
measurement with dephasing noise. The result density matrix in this
case should be unique. When False, the result will be conditioned on
sampled (but unreported) measurement results. In this case the result
may vary from call to call.}
}
\value{
The density matrix for the state which results from
applying the given operations to the desired initial state.
}
\description{
Returns the density matrix resulting from
simulating the circuit.
}
\details{
Note that, unlike `cirq.final_wavefunction`, terminal measurements
are not omitted. Instead, all measurements are treated as sources
of decoherence (i.e. measurements do not collapse, they dephase). See
ignore_measurement_results for details.
}
\seealso{
Other Sampling, Simulations and Data collection: 
\code{\link{big_endian_bits_to_int}()},
\code{\link{big_endian_digits_to_int}()},
\code{\link{big_endian_int_to_bits}()},
\code{\link{big_endian_int_to_digits}()},
\code{\link{circuit_like}()},
\code{\link{final_wavefunction}()},
\code{\link{flatten_to_ops_or_moments}()},
\code{\link{flatten_to_ops}()},
\code{\link{flatten_with_params}()},
\code{\link{flatten_with_sweep}()},
\code{\link{flatten}()},
\code{\link{hog_score_xeb_fidelity_from_probabilities}()},
\code{\link{measure_density_matrix}()},
\code{\link{measure_state_vector}()},
\code{\link{random_state_or_seed_like}()},
\code{\link{sample_density_matrix}()},
\code{\link{sample_state_vector}()},
\code{\link{sample_sweep}()},
\code{\link{sample}()},
\code{\link{to_resolvers}()}
}
\concept{Sampling, Simulations and Data collection}
