% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_simulations.R
\name{final_wavefunction}
\alias{final_wavefunction}
\title{Final wavefunction}
\usage{
final_wavefunction(program, ...)
}
\arguments{
\item{program}{The circuit, gate, operation, or tree of
operations to apply to the initial state in order to
produce the result.}

\item{...}{additional parameters to pass.}
}
\value{
The wavefunction resulting from applying the given
unitary operations to the desired initial state. Specifically,
a numpy array containing the the amplitudes in np.kron order,
where the order of arguments to kron is determined by the qubit
order argument (which defaults to just sorting the qubits that
are present into an ascending order).
}
\description{
Returns the state vector resulting from
acting operations on a state.
}
\details{
By default the input state is the computational
basis zero state, in which
case the output is just the first column of the implied
unitary matrix.
}
\seealso{
Other Sampling, Simulations and Data collection: 
\code{\link{big_endian_bits_to_int}()},
\code{\link{big_endian_digits_to_int}()},
\code{\link{big_endian_int_to_bits}()},
\code{\link{big_endian_int_to_digits}()},
\code{\link{circuit_like}()},
\code{\link{final_density_matrix}()},
\code{\link{flatten_to_ops_or_moments}()},
\code{\link{flatten_to_ops}()},
\code{\link{flatten_with_params}()},
\code{\link{flatten_with_sweep}()},
\code{\link{flatten}()},
\code{\link{hog_score_xeb_fidelity_from_probabilities}()},
\code{\link{measure_density_matrix}()},
\code{\link{measure_state_vector}()},
\code{\link{random_state_or_seed_like}()},
\code{\link{sample}()}
}
\concept{Sampling, Simulations and Data collection}
