% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_simulations.R
\name{flatten}
\alias{flatten}
\title{Flatten}
\usage{
flatten(val, ...)
}
\arguments{
\item{val}{The value to copy and substitute parameter expressions with flattened symbols.}

\item{...}{additional parameters to pass.}
}
\value{
The list (new value, expression map) where new value and expression map are described above.
}
\description{
Creates a copy of `val` with any symbols or
expressions replaced with new symbols. `val` can be a `Circuit`, `Gate`, `Operation`, or other
type.
}
\details{
`flatten` goes through every parameter in `val` and does the following:
- If the parameter is a number, don't change it.
- If the parameter is a symbol, don't change it.
- If the parameter is an expression, replace it with a symbol. The new symbol will be
`sympy.Symbol('<x + 1>')` if the expression was `sympy.Symbol('x') + 1`.
In the unlikely case that an expression with a different meaning also has the
string `'x + 1'`, a number is appended to the name to avoid collision:
`sympy.Symbol('<x + 1>_1')`. This function also creates a dictionary
mapping from expressions and symbols
in `val` to the new symbols in the flattened copy of `val`. E.g
`cirq.ExpressionMap({sympy.Symbol('x')+1: sympy.Symbol('<x + 1>')})`. This
`ExpressionMap` can be used to transform a sweep over the symbols in `val`
to a sweep over the flattened symbols e.g. a sweep over `sympy.Symbol('x')`
to a sweep over `sympy.Symbol('<x + 1>')`.
}
\seealso{
Other Sampling, Simulations and Data collection: 
\code{\link{big_endian_bits_to_int}()},
\code{\link{big_endian_digits_to_int}()},
\code{\link{big_endian_int_to_bits}()},
\code{\link{big_endian_int_to_digits}()},
\code{\link{circuit_like}()},
\code{\link{final_density_matrix}()},
\code{\link{final_wavefunction}()},
\code{\link{flatten_to_ops_or_moments}()},
\code{\link{flatten_to_ops}()},
\code{\link{flatten_with_params}()},
\code{\link{flatten_with_sweep}()},
\code{\link{hog_score_xeb_fidelity_from_probabilities}()},
\code{\link{measure_density_matrix}()},
\code{\link{measure_state_vector}()},
\code{\link{random_state_or_seed_like}()},
\code{\link{sample}()}
}
\concept{Sampling, Simulations and Data collection}
