% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noisy_Gates.R
\name{gate_asymmetric_depolarize}
\alias{gate_asymmetric_depolarize}
\title{Asymmetric depolarize}
\usage{
gate_asymmetric_depolarize(p_x, p_y, p_z)
}
\arguments{
\item{p_x}{The probability that a Pauli X and no other gate occurs.}

\item{p_y}{The probability that a Pauli Y and no other gate occurs.}

\item{p_z}{The probability that a Pauli Z and no other gate occurs.}
}
\value{
a AsymmetricDepolarizingChannel with given parameter.
}
\description{
This channel evolves a density matrix via
$$ \\rho \\rightarrow (1 - p_x - p_y - p_z) \\rho + p_x X \\rho X + p_y Y \\rho Y + p_z Z \\rho Z $$
}
\section{Raises}{

ValueError: if the args or the sum of the args are not probabilities.
}

\seealso{
Other Noisy gates and operations: 
\code{\link{gate_amplitude_damping_channel}()},
\code{\link{gate_amplitude_damp}()},
\code{\link{gate_asymmetric_depolarizing_channel}()},
\code{\link{gate_bit_flip_channel}()},
\code{\link{gate_bit_flip}()},
\code{\link{gate_depolarize}()},
\code{\link{gate_depolarizing_channel}()},
\code{\link{gate_generalized_amplitude_damping_channel}()},
\code{\link{gate_generalized_amplitude_damp}()},
\code{\link{gate_no_noise}()},
\code{\link{gate_noise_model_like}()},
\code{\link{gate_noise_model}()},
\code{\link{gate_phase_damping_channel}()},
\code{\link{gate_phase_damp}()},
\code{\link{gate_phase_flip_channel}()},
\code{\link{gate_phase_flip}()},
\code{\link{gate_reset_channel}()},
\code{\link{gate_reset}()}
}
\concept{Noisy gates and operations}
