% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gates.R
\name{gate_ccz}
\alias{gate_ccz}
\title{CCZ}
\usage{
gate_ccz(...)
}
\arguments{
\item{...}{qubits to pass. Expects 3 qubits.}
}
\value{
applied effects to a set of qubits.
}
\description{
the CCZ (doubly-controlled-Z) gate that can be raised to a power.
The exponent=1 instance of gate_ccx_pow.
}
\seealso{
Other Unitary gates and operations: 
\code{\link{gate_ccnot}()},
\code{\link{gate_ccx}()},
\code{\link{gate_cnot}()},
\code{\link{gate_cx}()},
\code{\link{gate_cz}()},
\code{\link{gate_h}()},
\code{\link{gate_i}()},
\code{\link{gate_rx}()},
\code{\link{gate_ry}()},
\code{\link{gate_s}()},
\code{\link{gate_t}()},
\code{\link{gate_xx}()},
\code{\link{gate_x}()},
\code{\link{gate_yy}()},
\code{\link{gate_y}()},
\code{\link{gate_zz}()},
\code{\link{gate_z}()}
}
\concept{Unitary gates and operations}
