% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unitary_Gates.R
\name{gate_ccz_pow}
\alias{gate_ccz_pow}
\title{CCZPowGate}
\usage{
gate_ccz_pow(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
A doubly-controlled-Z that can be raised to a power.
The matrix of `CCZ**t` is `diag(1, 1, 1, 1, 1, 1, 1, exp(i pi t))`.
}
\seealso{
Other Unitary gates and operations: 
\code{\link{gate_Gate}()},
\code{\link{gate_ccnot_pow}()},
\code{\link{gate_ccnot}()},
\code{\link{gate_ccx_pow}()},
\code{\link{gate_ccx}()},
\code{\link{gate_ccz}()},
\code{\link{gate_cnot_pow}()},
\code{\link{gate_cnot}()},
\code{\link{gate_controlled}()},
\code{\link{gate_cs_wap}()},
\code{\link{gate_cx_pow}()},
\code{\link{gate_cx}()},
\code{\link{gate_cz_pow}()},
\code{\link{gate_cz}()},
\code{\link{gate_eigen}()},
\code{\link{gate_fredkin}()},
\code{\link{gate_fsim}()},
\code{\link{gate_givens}()},
\code{\link{gate_global_phase_operation}()},
\code{\link{gate_hpow}()},
\code{\link{gate_h}()},
\code{\link{gate_identity_each}()},
\code{\link{gate_identity}()},
\code{\link{gate_is_wap_pow}()},
\code{\link{gate_iswap}()},
\code{\link{gate_i}()},
\code{\link{gate_matrix}()},
\code{\link{gate_operation}()},
\code{\link{gate_phase_gradient}()},
\code{\link{gate_phased_is_wap_pow}()},
\code{\link{gate_phased_xz}()},
\code{\link{gate_quantum_fourier_transform}()},
\code{\link{gate_riswap}()},
\code{\link{gate_rx}()},
\code{\link{gate_ry}()},
\code{\link{gate_single_qubit}()},
\code{\link{gate_swap_pow}()},
\code{\link{gate_swap}()},
\code{\link{gate_s}()},
\code{\link{gate_toffoli}()},
\code{\link{gate_t}()},
\code{\link{gate_xx}()},
\code{\link{gate_x}()},
\code{\link{gate_yy}()},
\code{\link{gate_y}()},
\code{\link{gate_zz}()},
\code{\link{gate_z}()}
}
\concept{Unitary gates and operations}
