% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noisy_Gates.R
\name{gate_depolarize}
\alias{gate_depolarize}
\title{Depolarize}
\usage{
gate_depolarize(p)
}
\arguments{
\item{p}{The probability that one of the Pauli gates is applied.
Each of the Pauli gates is applied independently with probability p / 3.}
}
\value{
a DepolarizingChannel with given probability of error.
}
\description{
This channel applies one of four disjoint possibilities: nothing (the
identity channel) or one of the three pauli gates. The disjoint
probabilities of the three gates are all the same, p / 3, and the
identity is done with probability 1 - p. The supplied probability
must be a valid probability or else this constructor will raise a
ValueError. This channel evolves a density matrix via
$$ \\rho \\rightarrow (1 - p) \\rho + (p / 3) X \\rho X + (p / 3) Y \\rho Y + (p / 3) Z \\rho Z $$
}
\section{This channel applies one of four disjoint possibilities}{
 nothing (the:
identity channel) or one of the three pauli gates. The disjoint probabilities of
the three gates are all the same, p / 3, and the identity is done with probability
1 - p. The supplied probability must be a valid probability or else this constructor
will raise a ValueError.
}

\section{Raises}{

ValueError: if p is not a valid probability.
}

\seealso{
Other Noisy gates and operations: 
\code{\link{gate_amplitude_damping_channel}()},
\code{\link{gate_amplitude_damp}()},
\code{\link{gate_asymmetric_depolarize}()},
\code{\link{gate_asymmetric_depolarizing_channel}()},
\code{\link{gate_bit_flip_channel}()},
\code{\link{gate_bit_flip}()},
\code{\link{gate_depolarizing_channel}()},
\code{\link{gate_generalized_amplitude_damping_channel}()},
\code{\link{gate_generalized_amplitude_damp}()},
\code{\link{gate_no_noise}()},
\code{\link{gate_noise_model_like}()},
\code{\link{gate_noise_model}()},
\code{\link{gate_phase_damping_channel}()},
\code{\link{gate_phase_damp}()},
\code{\link{gate_phase_flip_channel}()},
\code{\link{gate_phase_flip}()},
\code{\link{gate_reset_channel}()},
\code{\link{gate_reset}()}
}
\concept{Noisy gates and operations}
