% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noisy_Gates.R
\name{gate_generalized_amplitude_damp}
\alias{gate_generalized_amplitude_damp}
\title{Generalized amplitude damp}
\usage{
gate_generalized_amplitude_damp(p, gamma)
}
\arguments{
\item{p}{the probability of the qubit and environment exchanging energy.}

\item{gamma}{the probability of the interaction being dissipative.}
}
\value{
a GeneralizedAmplitudeDampingChannel with the given
}
\description{
probabilities gamma and p. This channel evolves a density matrix via:
$$ \\rho \\rightarrow M_0 \\rho M_0^\\dagger + M_1 \\rho M_1^\\dagger + M_2 \\rho M_2^\\dagger + M_3 \\rho M_3^\\dagger $$
With: $$ \\begin{aligned} M_0 =& \\sqrt{p} \\begin{bmatrix} 1 & 0 \\ 0 & \\sqrt{1 - \\gamma} \\end{bmatrix} \\ M_1 =& \\sqrt{p} \\begin{bmatrix} 0 & \\sqrt{\\gamma} \\ 0 & 0 \\end{bmatrix} \\ M_2 =& \\sqrt{1-p} \\begin{bmatrix} \\sqrt{1-\\gamma} & 0 \\ 0 & 1 \\end{bmatrix} \\ M_3 =& \\sqrt{1-p} \\begin{bmatrix} 0 & 0 \\ \\sqrt{\\gamma} & 0 \\end{bmatrix} \\end{aligned} $$
}
\section{Raises}{

ValueError: gamma or p is not a valid probability.
}

\seealso{
Other Noisy gates and operations: 
\code{\link{gate_amplitude_damping_channel}()},
\code{\link{gate_amplitude_damp}()},
\code{\link{gate_asymmetric_depolarize}()},
\code{\link{gate_asymmetric_depolarizing_channel}()},
\code{\link{gate_bit_flip_channel}()},
\code{\link{gate_bit_flip}()},
\code{\link{gate_depolarize}()},
\code{\link{gate_depolarizing_channel}()},
\code{\link{gate_generalized_amplitude_damping_channel}()},
\code{\link{gate_no_noise}()},
\code{\link{gate_noise_model_like}()},
\code{\link{gate_noise_model}()},
\code{\link{gate_phase_damping_channel}()},
\code{\link{gate_phase_damp}()},
\code{\link{gate_phase_flip_channel}()},
\code{\link{gate_phase_flip}()},
\code{\link{gate_reset_channel}()},
\code{\link{gate_reset}()}
}
\concept{Noisy gates and operations}
