% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noisy_Gates.R
\name{gate_generalized_amplitude_damping_channel}
\alias{gate_generalized_amplitude_damping_channel}
\title{Generalized Amplitude Damping Channel}
\usage{
gate_generalized_amplitude_damping_channel(p, gamma)
}
\arguments{
\item{p}{probability}

\item{gamma}{gamma}
}
\value{
The generalized amplitude damping channel object.
}
\description{
Dampen qubit amplitudes through non ideal dissipation.
}
\details{
This channel models the effect of energy dissipation into the environment
as well as the environment depositing energy into the system.
}
\seealso{
Other Noisy gates and operations: 
\code{\link{gate_amplitude_damping_channel}()},
\code{\link{gate_amplitude_damp}()},
\code{\link{gate_asymmetric_depolarize}()},
\code{\link{gate_asymmetric_depolarizing_channel}()},
\code{\link{gate_bit_flip_channel}()},
\code{\link{gate_bit_flip}()},
\code{\link{gate_depolarize}()},
\code{\link{gate_depolarizing_channel}()},
\code{\link{gate_generalized_amplitude_damp}()},
\code{\link{gate_no_noise}()},
\code{\link{gate_noise_model_like}()},
\code{\link{gate_noise_model}()},
\code{\link{gate_phase_damping_channel}()},
\code{\link{gate_phase_damp}()},
\code{\link{gate_phase_flip_channel}()},
\code{\link{gate_phase_flip}()},
\code{\link{gate_reset_channel}()},
\code{\link{gate_reset}()}
}
\concept{Noisy gates and operations}
