% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unitary_Gates.R
\name{gate_givens}
\alias{gate_givens}
\title{Givens}
\usage{
gate_givens(angle_rads)
}
\arguments{
\item{angle_rads}{The rotation angle in radians.}
}
\value{
A phased iswap gate for the given rotation.
}
\description{
a gate with matrix exp(-i angle_rads (Y⊗X - X⊗Y) / 2).
}
\details{
In numerical linear algebra Givens rotation is any linear transformation
with matrix equal to the identity except for a 2x2 orthogonal submatrix
[[cos(a), -sin(a)],
[sin(a), cos(a)]] which performs a 2D rotation on a
subspace spanned by two basis vectors. In quantum computational chemistry
the term is used to refer to the two-qubit gate defined as
givens(a) ≡ exp(-i a (Y⊗X - X⊗Y) / 2) with the matrix
[[1, 0, 0, 0],
[0, c, -s, 0],
[0, s, c, 0],
[0, 0, 0, 1]] where
c = cos(a), s = sin(a). The matrix is a Givens rotation in the numerical
linear algebra sense acting on the subspace spanned by the |01⟩ and |10⟩ states.
The gate is also equivalent to the ISWAP conjugated by T^-1 ⊗ T.
}
\seealso{
Other Unitary gates and operations: 
\code{\link{gate_ccnot_pow}()},
\code{\link{gate_ccnot}()},
\code{\link{gate_ccx}()},
\code{\link{gate_ccz}()},
\code{\link{gate_cnot}()},
\code{\link{gate_cx}()},
\code{\link{gate_cz}()},
\code{\link{gate_fredkin}()},
\code{\link{gate_h}()},
\code{\link{gate_identity_each}()},
\code{\link{gate_iswap}()},
\code{\link{gate_i}()},
\code{\link{gate_riswap}()},
\code{\link{gate_rx}()},
\code{\link{gate_ry}()},
\code{\link{gate_swap}()},
\code{\link{gate_s}()},
\code{\link{gate_toffoli}()},
\code{\link{gate_t}()},
\code{\link{gate_xx}()},
\code{\link{gate_x}()},
\code{\link{gate_yy}()},
\code{\link{gate_y}()},
\code{\link{gate_zz}()},
\code{\link{gate_z}()}
}
\concept{Unitary gates and operations}
