% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noisy_Gates.R
\name{gate_noise_model}
\alias{gate_noise_model}
\title{Noise Model}
\usage{
gate_noise_model(...)
}
\arguments{
\item{...}{parameters to pass.}
}
\value{
The generalized amplitude damping channel object.
}
\description{
Replaces operations and moments with noisy counterparts.
}
\details{
A child class must override at least one of the following three methods:
- noisy_moments
- noisy_moment
- noisy_operation
The methods that are not overridden will be implemented in terms of the ones
that are.
Simulators told to use a noise model will use these methods in order to
dynamically rewrite the program they are simulating.
}
\seealso{
Other Noisy gates and operations: 
\code{\link{gate_amplitude_damping_channel}()},
\code{\link{gate_amplitude_damp}()},
\code{\link{gate_asymmetric_depolarize}()},
\code{\link{gate_asymmetric_depolarizing_channel}()},
\code{\link{gate_bit_flip_channel}()},
\code{\link{gate_bit_flip}()},
\code{\link{gate_depolarize}()},
\code{\link{gate_depolarizing_channel}()},
\code{\link{gate_generalized_amplitude_damping_channel}()},
\code{\link{gate_generalized_amplitude_damp}()},
\code{\link{gate_no_noise}()},
\code{\link{gate_noise_model_like}()},
\code{\link{gate_phase_damping_channel}()},
\code{\link{gate_phase_damp}()},
\code{\link{gate_phase_flip_channel}()},
\code{\link{gate_phase_flip}()},
\code{\link{gate_reset_channel}()},
\code{\link{gate_reset}()}
}
\concept{Noisy gates and operations}
