% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noisy_Gates.R
\name{gate_noise_model_like}
\alias{gate_noise_model_like}
\title{Noise model like}
\usage{
gate_noise_model_like(...)
}
\arguments{
\item{...}{parameters to pass.}
}
\value{
None
}
\description{
A cirq.NoiseModel or a value that can be
trivially converted into one.
None is a NOISE_MODEL_LIKE. It will be replaced by the
cirq.NO_NOISE noise model.
A single qubit gate is a NOISE_MODEL_LIKE. It will be
wrapped inside of a cirq.ConstantQubitNoiseModel.
}
\seealso{
Other Noisy gates and operations: 
\code{\link{gate_amplitude_damping_channel}()},
\code{\link{gate_amplitude_damp}()},
\code{\link{gate_asymmetric_depolarize}()},
\code{\link{gate_asymmetric_depolarizing_channel}()},
\code{\link{gate_bit_flip_channel}()},
\code{\link{gate_bit_flip}()},
\code{\link{gate_depolarize}()},
\code{\link{gate_depolarizing_channel}()},
\code{\link{gate_generalized_amplitude_damping_channel}()},
\code{\link{gate_generalized_amplitude_damp}()},
\code{\link{gate_no_noise}()},
\code{\link{gate_noise_model}()},
\code{\link{gate_phase_damping_channel}()},
\code{\link{gate_phase_damp}()},
\code{\link{gate_phase_flip_channel}()},
\code{\link{gate_phase_flip}()},
\code{\link{gate_reset_channel}()},
\code{\link{gate_reset}()}
}
\concept{Noisy gates and operations}
