% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circuits.R
\name{gate_operation_}
\alias{gate_operation_}
\title{Gate Operation}
\usage{
gate_operation_(gate, qubits)
}
\arguments{
\item{gate}{the gate to apply.}

\item{qubits}{the qubits to operate on.}
}
\value{
operation applied to the sequence of qubits.
}
\description{
An application of a gate to a sequence of qubits.
}
\seealso{
Other Circuits: 
\code{\link{circuit_dag}()},
\code{\link{circuit}()},
\code{\link{gate_parallel_operation}()},
\code{\link{insert_strategy}()},
\code{\link{moment}()},
\code{\link{qubit_order_or_list}()},
\code{\link{qubit_order}()},
\code{\link{tree_flatten_op}()},
\code{\link{tree_op}()}
}
\concept{Circuits}
