% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circuits.R
\name{gate_parallel_operation}
\alias{gate_parallel_operation}
\title{Parallel Gate Operation}
\usage{
gate_parallel_operation(...)
}
\arguments{
\item{...}{parameters to pass.}
}
\value{
applied operations to the qubits.
}
\description{
An application of several copies of a gate to a group of qubits.
}
\seealso{
Other Circuits: 
\code{\link{circuit_dag}()},
\code{\link{circuit}()},
\code{\link{gate_operation_}()},
\code{\link{insert_strategy}()},
\code{\link{moment}()},
\code{\link{qubit_order_or_list}()},
\code{\link{qubit_order}()},
\code{\link{tree_flatten_op}()},
\code{\link{tree_op}()}
}
\concept{Circuits}
