% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noisy_Gates.R
\name{gate_phase_damp}
\alias{gate_phase_damp}
\title{Phase damp}
\usage{
gate_phase_damp(gamma)
}
\arguments{
\item{gamma}{The damping constant.}
}
\value{
Creates a PhaseDampingChannel with damping constant gamma.
}
\description{
This channel evolves a density matrix via:
$$ \\rho \\rightarrow M_0 \\rho M_0^\\dagger + M_1 \\rho M_1^\\dagger $$
With: $$ \\begin{aligned} M_0 =& \\begin{bmatrix} 1 & 0 \\ 0 & \\sqrt{1 - \\gamma} \\end{bmatrix} \\ M_1 =& \\begin{bmatrix} 0 & 0 \\ 0 & \\sqrt{\\gamma} \\end{bmatrix} \\end{aligned} $$
}
\section{Raises}{

ValueError: is gamma is not a valid probability.
}

\seealso{
Other Noisy gates and operations: 
\code{\link{gate_amplitude_damping_channel}()},
\code{\link{gate_amplitude_damp}()},
\code{\link{gate_asymmetric_depolarize}()},
\code{\link{gate_asymmetric_depolarizing_channel}()},
\code{\link{gate_bit_flip_channel}()},
\code{\link{gate_bit_flip}()},
\code{\link{gate_depolarize}()},
\code{\link{gate_depolarizing_channel}()},
\code{\link{gate_generalized_amplitude_damping_channel}()},
\code{\link{gate_generalized_amplitude_damp}()},
\code{\link{gate_no_noise}()},
\code{\link{gate_noise_model_like}()},
\code{\link{gate_noise_model}()},
\code{\link{gate_phase_damping_channel}()},
\code{\link{gate_phase_flip_channel}()},
\code{\link{gate_phase_flip}()},
\code{\link{gate_reset_channel}()},
\code{\link{gate_reset}()}
}
\concept{Noisy gates and operations}
