% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noisy_Gates.R
\name{gate_phase_damping_channel}
\alias{gate_phase_damping_channel}
\title{Phase Damping Channel}
\usage{
gate_phase_damping_channel(gamma)
}
\arguments{
\item{gamma}{gamma}
}
\value{
The phase damping channel channel object.
}
\description{
Dampen qubit phase.
}
\details{
This channel models phase damping which is the loss of quantum
information without the loss of energy.
}
\seealso{
Other Noisy gates and operations: 
\code{\link{gate_amplitude_damping_channel}()},
\code{\link{gate_amplitude_damp}()},
\code{\link{gate_asymmetric_depolarize}()},
\code{\link{gate_asymmetric_depolarizing_channel}()},
\code{\link{gate_bit_flip_channel}()},
\code{\link{gate_bit_flip}()},
\code{\link{gate_depolarize}()},
\code{\link{gate_depolarizing_channel}()},
\code{\link{gate_generalized_amplitude_damping_channel}()},
\code{\link{gate_generalized_amplitude_damp}()},
\code{\link{gate_no_noise}()},
\code{\link{gate_noise_model_like}()},
\code{\link{gate_noise_model}()},
\code{\link{gate_phase_damp}()},
\code{\link{gate_phase_flip_channel}()},
\code{\link{gate_phase_flip}()},
\code{\link{gate_reset_channel}()},
\code{\link{gate_reset}()}
}
\concept{Noisy gates and operations}
