% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noisy_Gates.R
\name{gate_phase_flip}
\alias{gate_phase_flip}
\title{Phase flip}
\usage{
gate_phase_flip(p = NULL)
}
\arguments{
\item{p}{the probability of a phase flip.}
}
\value{
a phase flip channel that flips a qubit's phase with probability p
}
\description{
if p is NULL, return a guaranteed phase flip in
the form of a Z operation. This channel evolves a density matrix
via: $$ \\rho \\rightarrow M_0 \\rho M_0^\\dagger + M_1 \\rho M_1^\\dagger $$
With: $$ \\begin{aligned} M_0 =& \\sqrt{p} \\begin{bmatrix} 1 & 0 \\ 0 & 1 \\end{bmatrix} \\ M_1 =& \\sqrt{1-p} \\begin{bmatrix} 1 & 0 \\ 0 & -1 \\end{bmatrix} \\end{aligned} $$
}
\section{Raises}{

ValueError: if p is not a valid probability.
}

\seealso{
Other Noisy gates and operations: 
\code{\link{gate_amplitude_damping_channel}()},
\code{\link{gate_amplitude_damp}()},
\code{\link{gate_asymmetric_depolarize}()},
\code{\link{gate_asymmetric_depolarizing_channel}()},
\code{\link{gate_bit_flip_channel}()},
\code{\link{gate_bit_flip}()},
\code{\link{gate_depolarize}()},
\code{\link{gate_depolarizing_channel}()},
\code{\link{gate_generalized_amplitude_damping_channel}()},
\code{\link{gate_generalized_amplitude_damp}()},
\code{\link{gate_no_noise}()},
\code{\link{gate_noise_model_like}()},
\code{\link{gate_noise_model}()},
\code{\link{gate_phase_damping_channel}()},
\code{\link{gate_phase_damp}()},
\code{\link{gate_phase_flip_channel}()},
\code{\link{gate_reset_channel}()},
\code{\link{gate_reset}()}
}
\concept{Noisy gates and operations}
