% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unitary_Gates.R
\name{gate_qft}
\alias{gate_qft}
\title{QFT}
\usage{
gate_qft(qubits, without_reverse = FALSE, inverse = FALSE)
}
\arguments{
\item{qubits}{– The qubits to apply the QFT to.}

\item{without_reverse}{When set, swap gates at the end of the
QFT are omitted. This reverses the qubit order relative to the
standard QFT effect, but makes the gate cheaper to apply.}

\item{inverse}{If set, the inverse QFT is performed instead
of the QFT. Equivalent to calling cirq.inverse on the result,
or raising it to the -1.}
}
\value{
A cirq.Operation applying the QFT to the given qubits.
}
\description{
The quantum Fourier transform. Transforms a qubit
register from the computational basis to the frequency basis.
The inverse quantum Fourier transform is `cirq.QFT(*qubits)**-1`
or equivalently `cirq.inverse(cirq.QFT(*qubits))`.
}
\concept{Unitary gates and operations.}
