% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Noisy_Gates.R
\name{gate_reset_channel}
\alias{gate_reset_channel}
\title{ResetChannel}
\usage{
gate_reset_channel(dimension = 2)
}
\arguments{
\item{dimension}{dimension}
}
\value{
The phase flip channel channel object.
}
\description{
Reset a qubit back to its |0⟩ state.
}
\details{
The reset channel is equivalent to performing an unobserved measurement
which then controls a bit flip onto the targeted qubit.
}
\seealso{
Other Noisy gates and operations: 
\code{\link{gate_amplitude_damping_channel}()},
\code{\link{gate_amplitude_damp}()},
\code{\link{gate_asymmetric_depolarize}()},
\code{\link{gate_asymmetric_depolarizing_channel}()},
\code{\link{gate_bit_flip_channel}()},
\code{\link{gate_bit_flip}()},
\code{\link{gate_depolarize}()},
\code{\link{gate_depolarizing_channel}()},
\code{\link{gate_generalized_amplitude_damping_channel}()},
\code{\link{gate_generalized_amplitude_damp}()},
\code{\link{gate_no_noise}()},
\code{\link{gate_noise_model_like}()},
\code{\link{gate_noise_model}()},
\code{\link{gate_phase_damping_channel}()},
\code{\link{gate_phase_damp}()},
\code{\link{gate_phase_flip_channel}()},
\code{\link{gate_phase_flip}()},
\code{\link{gate_reset}()}
}
\concept{Noisy gates and operations}
