% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gates.R
\name{gate_rx}
\alias{gate_rx}
\title{RX}
\usage{
gate_rx(rads)
}
\arguments{
\item{rads}{rads}
}
\value{
Returns a gate with the matrix e^{-i X rads / 2}.
}
\description{
the controlled RX gate.
}
\seealso{
Other Unitary gates and operations: 
\code{\link{gate_ccnot}()},
\code{\link{gate_ccx}()},
\code{\link{gate_ccz}()},
\code{\link{gate_cnot}()},
\code{\link{gate_cx}()},
\code{\link{gate_cz}()},
\code{\link{gate_h}()},
\code{\link{gate_i}()},
\code{\link{gate_ry}()},
\code{\link{gate_s}()},
\code{\link{gate_t}()},
\code{\link{gate_xx}()},
\code{\link{gate_x}()},
\code{\link{gate_yy}()},
\code{\link{gate_y}()},
\code{\link{gate_zz}()},
\code{\link{gate_z}()}
}
\concept{Unitary gates and operations}
