% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unitary_Gates.R
\name{gate_ry}
\alias{gate_ry}
\title{RY}
\usage{
gate_ry(rads)

gate_ry(rads)
}
\arguments{
\item{rads}{rads}
}
\value{
Returns a gate with the matrix e^{-i Y rads / 2}.

a gate with the matrix e^{-i Z rads / 2}.
}
\description{
the controlled RY gate.

the controlled RZ gate.
}
\seealso{
Other Unitary gates and operations: 
\code{\link{gate_ccnot_pow}()},
\code{\link{gate_ccnot}()},
\code{\link{gate_ccx}()},
\code{\link{gate_ccz}()},
\code{\link{gate_cnot}()},
\code{\link{gate_cx}()},
\code{\link{gate_cz}()},
\code{\link{gate_fredkin}()},
\code{\link{gate_givens}()},
\code{\link{gate_h}()},
\code{\link{gate_identity_each}()},
\code{\link{gate_iswap}()},
\code{\link{gate_i}()},
\code{\link{gate_riswap}()},
\code{\link{gate_rx}()},
\code{\link{gate_swap}()},
\code{\link{gate_s}()},
\code{\link{gate_toffoli}()},
\code{\link{gate_t}()},
\code{\link{gate_xx}()},
\code{\link{gate_x}()},
\code{\link{gate_yy}()},
\code{\link{gate_y}()},
\code{\link{gate_zz}()},
\code{\link{gate_z}()}

Other Unitary gates and operations: 
\code{\link{gate_ccnot_pow}()},
\code{\link{gate_ccnot}()},
\code{\link{gate_ccx}()},
\code{\link{gate_ccz}()},
\code{\link{gate_cnot}()},
\code{\link{gate_cx}()},
\code{\link{gate_cz}()},
\code{\link{gate_fredkin}()},
\code{\link{gate_givens}()},
\code{\link{gate_h}()},
\code{\link{gate_identity_each}()},
\code{\link{gate_iswap}()},
\code{\link{gate_i}()},
\code{\link{gate_riswap}()},
\code{\link{gate_rx}()},
\code{\link{gate_swap}()},
\code{\link{gate_s}()},
\code{\link{gate_toffoli}()},
\code{\link{gate_t}()},
\code{\link{gate_xx}()},
\code{\link{gate_x}()},
\code{\link{gate_yy}()},
\code{\link{gate_y}()},
\code{\link{gate_zz}()},
\code{\link{gate_z}()}
}
\concept{Unitary gates and operations}
