% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Unitary_Gates.R
\name{gate_toffoli}
\alias{gate_toffoli}
\title{TOFFOLI}
\usage{
gate_toffoli(...)
}
\arguments{
\item{...}{qubits to pass. Expects 3 qubits.}
}
\value{
applied effects to a set of qubits.
}
\description{
the tofolli gate. The exponent=1 instance of gate_ccx_pow.
}
\seealso{
Other Unitary gates and operations: 
\code{\link{gate_ccnot_pow}()},
\code{\link{gate_ccnot}()},
\code{\link{gate_ccx}()},
\code{\link{gate_ccz}()},
\code{\link{gate_cnot}()},
\code{\link{gate_cx}()},
\code{\link{gate_cz}()},
\code{\link{gate_fredkin}()},
\code{\link{gate_givens}()},
\code{\link{gate_h}()},
\code{\link{gate_identity_each}()},
\code{\link{gate_iswap}()},
\code{\link{gate_i}()},
\code{\link{gate_riswap}()},
\code{\link{gate_rx}()},
\code{\link{gate_ry}()},
\code{\link{gate_swap}()},
\code{\link{gate_s}()},
\code{\link{gate_t}()},
\code{\link{gate_xx}()},
\code{\link{gate_x}()},
\code{\link{gate_yy}()},
\code{\link{gate_y}()},
\code{\link{gate_zz}()},
\code{\link{gate_z}()}
}
\concept{Unitary gates and operations}
