% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Gates.R
\name{gate_x}
\alias{gate_x}
\title{Pauli X}
\usage{
gate_x(...)
}
\arguments{
\item{...}{qubits to pass. Expects 2 qubits.}
}
\value{
applied effects to a set of qubits.
}
\description{
the Pauli X gate.
}
\seealso{
Other Unitary gates and operations: 
\code{\link{gate_ccnot}()},
\code{\link{gate_ccx}()},
\code{\link{gate_ccz}()},
\code{\link{gate_cnot}()},
\code{\link{gate_cx}()},
\code{\link{gate_cz}()},
\code{\link{gate_h}()},
\code{\link{gate_i}()},
\code{\link{gate_rx}()},
\code{\link{gate_ry}()},
\code{\link{gate_s}()},
\code{\link{gate_t}()},
\code{\link{gate_xx}()},
\code{\link{gate_yy}()},
\code{\link{gate_y}()},
\code{\link{gate_zz}()},
\code{\link{gate_z}()}
}
\concept{Unitary gates and operations}
