% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Devices.R
\name{grid_qubit}
\alias{grid_qubit}
\title{Grid Qubit}
\usage{
grid_qubit(row, col)
}
\arguments{
\item{row}{row}

\item{col}{col}
}
\value{
grid qubit
}
\description{
A qubit on a 2d square lattice.
}
\details{
GridQubits use row-major ordering:
GridQubit(0, 0) < GridQubit(0, 1) < GridQubit(1, 0) < GridQubit(1, 1)
New GridQubits can be constructed by adding or subtracting lists
grid_qubit(2, 3) + (3, 1) grid_qubit(5, 4)
grid_qubit(2, 3) - (1, 2) grid_qubit(1, 1)
}
\seealso{
Other Devices: 
\code{\link{grid_qid}()},
\code{\link{line_qid}()},
\code{\link{line_qubit}()},
\code{\link{named_qubit}()},
\code{\link{qid}()},
\code{\link{unconstrained_device}()}
}
\concept{Devices}
