% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_simulations.R
\name{hog_score_xeb_fidelity_from_probabilities}
\alias{hog_score_xeb_fidelity_from_probabilities}
\title{Hog score xeb fidelity from probabilities}
\usage{
hog_score_xeb_fidelity_from_probabilities(
  hilbert_space_dimension,
  probabilities
)
}
\arguments{
\item{hilbert_space_dimension}{Dimension of the Hilbert space on
which the channel whose fidelity is being estimated is defined.}

\item{probabilities}{Ideal probabilities of bitstrings observed
in experiment. Returns: Estimate of fidelity associated with an
experimental realization of a quantum circuit.}
}
\value{
Estimate of fidelity associated with an experimental realization of a quantum circuit.
}
\description{
XEB fidelity estimator based on normalized HOG score.
}
\details{
Estimates fidelity from ideal probabilities of observed bitstrings.
See `linear_xeb_fidelity_from_probabilities` for the assumptions made
by this estimator. The mean of this estimator is the true fidelity f and the
variance is (1/log(2)^2 - f^2) / M where f is the fidelity and M the number
of observations, equal to
len(probabilities). This is always worse than log XEB (see above).
Since this estimator is unbiased, the variance is equal to the mean
squared error of the estimator. The estimator is intended for use with xeb_fidelity() below. It is
based on the HOG problem defined in https://arxiv.org/abs/1612.05903.
}
\seealso{
Other Sampling, Simulations and Data collection: 
\code{\link{big_endian_bits_to_int}()},
\code{\link{big_endian_digits_to_int}()},
\code{\link{big_endian_int_to_bits}()},
\code{\link{big_endian_int_to_digits}()},
\code{\link{circuit_like}()},
\code{\link{final_density_matrix}()},
\code{\link{final_wavefunction}()},
\code{\link{flatten_to_ops_or_moments}()},
\code{\link{flatten_to_ops}()},
\code{\link{flatten_with_params}()},
\code{\link{flatten_with_sweep}()},
\code{\link{flatten}()},
\code{\link{measure_density_matrix}()},
\code{\link{measure_state_vector}()},
\code{\link{random_state_or_seed_like}()},
\code{\link{sample_density_matrix}()},
\code{\link{sample_state_vector}()},
\code{\link{sample_sweep}()},
\code{\link{sample}()},
\code{\link{to_resolvers}()}
}
\concept{Sampling, Simulations and Data collection}
