% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circuits.R
\name{insert_strategy}
\alias{insert_strategy}
\title{Insert Strategy}
\usage{
insert_strategy(...)
}
\arguments{
\item{...}{parameters to pass.}
}
\value{
added operation to a circuit
}
\description{
Indicates preferences on how to add multiple
operations to a circuit.
}
\seealso{
Other Circuits: 
\code{\link{circuit_dag}()},
\code{\link{circuit}()},
\code{\link{gate_operation_}()},
\code{\link{gate_parallel_operation}()},
\code{\link{moment}()},
\code{\link{qubit_order_or_list}()},
\code{\link{qubit_order}()},
\code{\link{tree_flatten_op}()},
\code{\link{tree_op}()}
}
\concept{Circuits}
