% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_cirq}
\alias{install_cirq}
\title{Install Cirq}
\usage{
install_cirq(version = NULL, ..., restart_session = TRUE, from_git = FALSE)
}
\arguments{
\item{version}{for specific version of Cirq, e.g. "0.8.0"}

\item{...}{other arguments passed to [reticulate::py_install()].}

\item{restart_session}{Restart R session after installing (note this will only occur within RStudio).}

\item{from_git}{install the recent GitHub version of Cirq}
}
\value{
a python module cirq
}
\description{
This function is used to install the Cirq python module
}
