% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Devices.R
\name{line_qid}
\alias{line_qid}
\title{Line Qid}
\usage{
line_qid(x, dimension)
}
\arguments{
\item{x}{The x coordinate.}

\item{dimension}{The dimension of the qid, e.g. the number of quantum levels.}
}
\value{
initialized line qid at the given x coordinate.
}
\description{
A qid on a 1d lattice with nearest-neighbor connectivity.
}
\details{
`LineQid`s have a single attribute, and integer coordinate 'x', which
identifies the qids location on the line. `LineQid`s are ordered by
this integer. One can construct new `LineQid`s by adding or subtracting integers:
line_qid(1, dimension=2) + 3
Output: cirq.LineQid(4, dimension=2)
line_qid(2, dimension=3) - 1
Output: cirq.LineQid(1, dimension=3)
}
\seealso{
Other Devices: 
\code{\link{grid_qid}()},
\code{\link{grid_qubit}()},
\code{\link{line_qubit}()},
\code{\link{named_qubit}()},
\code{\link{qid}()},
\code{\link{unconstrained_device}()}
}
\concept{Devices}
