% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Devices.R
\name{line_qubit}
\alias{line_qubit}
\title{Line Qubit}
\usage{
line_qubit(x)
}
\arguments{
\item{x}{integer coordinate}
}
\value{
initialized line qubit at the given x coordinate.
}
\description{
A qubit on a 1d lattice with nearest-neighbor connectivity.
}
\details{
LineQubits have a single attribute, and integer coordinate 'x', which
identifies the qubits location on the line. LineQubits are ordered by
this integer. One can construct new LineQubits by adding or subtracting integers:
line_qubitQubit(1) + 3
Output: cirq.LineQubit(4)
line_qubit(2) - 1
Output: cirq.LineQubit(1)
}
\seealso{
Other Devices: 
\code{\link{grid_qid}()},
\code{\link{grid_qubit}()},
\code{\link{line_qid}()},
\code{\link{named_qubit}()},
\code{\link{qid}()},
\code{\link{unconstrained_device}()}
}
\concept{Devices}
