% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure.R
\name{measure}
\alias{measure}
\title{Measure}
\usage{
measure(...)
}
\arguments{
\item{...}{parameters to pass:
target – The qubits that the measurement gate should measure.

key – The string key of the measurement. If this is NULL, it
defaults to a comma-separated list of the target qubits’ str values.

invert_mask – A list of Truthy or Falsey values indicating
whether the corresponding qubits should be flipped. NULL
indicates no inverting should be done.}
}
\value{
An operation targeting the given qubits with a measurement.
}
\description{
a single MeasurementGate applied to all the given qubits.
The qubits are measured in the computational basis.
}
\note{
ValueError if the qubits are not instances of Qid.
}
\seealso{
Other Measurement: 
\code{\link{measure_each}()},
\code{\link{measurement_gate}()}
}
\concept{Measurement}
