% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Measure.R
\name{measure_each}
\alias{measure_each}
\title{Measure Each}
\usage{
measure_each(...)
}
\arguments{
\item{...}{parameters to pass:
*qubits – The qubits to measure.
key_func – Determines the key of the measurements of each qubit.
Takes the qubit and returns the key for that qubit. Defaults to str.}
}
\value{
a list of operations individually measuring the given qubits.
}
\description{
The qubits are measured in the computational basis.
}
\seealso{
Other Measurement: 
\code{\link{measurement_gate}()},
\code{\link{measure}()}
}
\concept{Measurement}
