% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Devices.R
\name{named_qubit}
\alias{named_qubit}
\title{Named Qubit}
\usage{
named_qubit(name)
}
\arguments{
\item{name}{name of the op.}
}
\value{
named qubit.
}
\description{
A qubit identified by name.
}
\details{
By default, NamedQubit has a lexicographic order. However, numbers within
the name are handled correctly. So, for example, if you print a circuit
containing `named_qubit('qubit22')` and `named_qubit('qubit3')`, the
wire for 'qubit3' will correctly come before 'qubit22'.
}
\seealso{
Other Devices: 
\code{\link{grid_qid}()},
\code{\link{grid_qubit}()},
\code{\link{line_qid}()},
\code{\link{line_qubit}()},
\code{\link{qid}()},
\code{\link{unconstrained_device}()}
}
\concept{Devices}
