% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pauli_Clifford.R
\name{pauli}
\alias{pauli}
\title{Pauli}
\usage{
pauli(index, name)
}
\arguments{
\item{index}{the index}

\item{name}{the name of op}
}
\value{
a pauli object.
}
\description{
Represents the Pauli gates.
}
\details{
This is an abstract class with no public subclasses. The only instances
of private subclasses are the X, Y, or Z Pauli gates defined below.
}
\seealso{
Other Pauli and Clifford groups: 
\code{\link{clifford_state}()},
\code{\link{clifford_tableau}()},
\code{\link{pauli_base_dense_string}()},
\code{\link{pauli_basis}()},
\code{\link{pauli_dense_string}()},
\code{\link{pauli_gate_like}()},
\code{\link{pauli_interaction_gate}()},
\code{\link{pauli_mutable_dense_string}()},
\code{\link{pauli_pow_combination}()}
}
\concept{Pauli and Clifford groups}
