% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pauli_Clifford.R
\name{pauli_base_dense_string}
\alias{pauli_base_dense_string}
\title{BaseDensePauliString}
\usage{
pauli_base_dense_string(pauli_mask, coefficient)
}
\arguments{
\item{pauli_mask}{A specification of the Pauli gates to use.
This argument can be a string like “IXYYZ”, or a numeric list
like [0, 1, 3, 2] with I=0, X=1, Y=2, Z=3=X|Y.
The internal representation is a 1-dimensional uint8 numpy array
containing numeric values. If such a numpy array is given, and the
pauli string is mutable, the argument will be used directly
instead of being copied.}

\item{coefficient}{A complex number. Usually +1, -1, 1j, or -1j but other values are supported.}
}
\description{
Parent class for `DensePauliString` and
`MutableDensePauliString`.
}
\seealso{
Other Pauli and Clifford groups: 
\code{\link{clifford_single_qubit_gate}()},
\code{\link{clifford_state}()},
\code{\link{clifford_tableau}()},
\code{\link{pauli_basis}()},
\code{\link{pauli_dense_string}()},
\code{\link{pauli_gate_like}()},
\code{\link{pauli_interaction_gate}()},
\code{\link{pauli_mutable_dense_string}()},
\code{\link{pauli_pow_combination}()},
\code{\link{pauli_string_gate_operation}()},
\code{\link{pauli_string}()},
\code{\link{pauli_sum_like}()},
\code{\link{pauli_sum}()},
\code{\link{pauli_transform}()},
\code{\link{pauli}()}
}
\concept{Pauli and Clifford groups}
