% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pauli_Clifford.R
\name{pauli_basis}
\alias{pauli_basis}
\title{PAULI BASIS}
\usage{
pauli_basis()
}
\value{
a named list with 4 matrices: I, X, Y, and Z.
}
\description{
The four Pauli matrices (including identity) keyed by character.
}
\seealso{
Other Pauli and Clifford groups: 
\code{\link{clifford_state}()},
\code{\link{clifford_tableau}()},
\code{\link{pauli_base_dense_string}()},
\code{\link{pauli_dense_string}()},
\code{\link{pauli_gate_like}()},
\code{\link{pauli_interaction_gate}()},
\code{\link{pauli_mutable_dense_string}()},
\code{\link{pauli_pow_combination}()},
\code{\link{pauli}()}
}
\concept{Pauli and Clifford groups}
