% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pauli_Clifford.R
\name{pauli_gate_like}
\alias{pauli_gate_like}
\title{PAULI GATE LIKE}
\usage{
pauli_gate_like(...)

pauli_gate_like(...)
}
\arguments{
\item{...}{parameters to pass.}
}
\value{
a pauli gate object.

Collections of Pauli operations are recrusively multiplied into a single Pauli string.
}
\description{
An object that can be interpreted as a Pauli gate.

A cirq.PauliString or a value that can easily be converted into one.
}
\details{
Allowed values are:
1. Cirq gates: cirq.I, cirq.X, cirq.Y, cirq.Z.
2. Strings: “I”, “X”, “Y”, “Z”. Equivalently “i”, “x”, “y”, “z”.
3. Integers from 0 to 3, with the convention 0=I, 1=X, 2=Y, 3=Z.

Dictionaries from qubit to Pauli operation are wrapped into a Pauli string.
Each Pauli operation can be specified as a cirq object (e.g. cirq.X) or as
a string (e.g. "X") or as an integer where 0=I, 1=X, 2=Y, 3=Z.
}
\seealso{
Other Pauli and Clifford groups: 
\code{\link{clifford_single_qubit_gate}()},
\code{\link{clifford_state}()},
\code{\link{clifford_tableau}()},
\code{\link{pauli_base_dense_string}()},
\code{\link{pauli_basis}()},
\code{\link{pauli_dense_string}()},
\code{\link{pauli_interaction_gate}()},
\code{\link{pauli_mutable_dense_string}()},
\code{\link{pauli_pow_combination}()},
\code{\link{pauli_string_gate_operation}()},
\code{\link{pauli_string}()},
\code{\link{pauli_sum_like}()},
\code{\link{pauli_sum}()},
\code{\link{pauli_transform}()},
\code{\link{pauli}()}

Other Pauli and Clifford groups: 
\code{\link{clifford_single_qubit_gate}()},
\code{\link{clifford_state}()},
\code{\link{clifford_tableau}()},
\code{\link{pauli_base_dense_string}()},
\code{\link{pauli_basis}()},
\code{\link{pauli_dense_string}()},
\code{\link{pauli_interaction_gate}()},
\code{\link{pauli_mutable_dense_string}()},
\code{\link{pauli_pow_combination}()},
\code{\link{pauli_string_gate_operation}()},
\code{\link{pauli_string}()},
\code{\link{pauli_sum_like}()},
\code{\link{pauli_sum}()},
\code{\link{pauli_transform}()},
\code{\link{pauli}()}
}
\concept{Pauli and Clifford groups}
