% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pauli_Clifford.R
\name{pauli_interaction_gate}
\alias{pauli_interaction_gate}
\title{Pauli Interaction Gate}
\usage{
pauli_interaction_gate(pauli0, invert0, pauli1, invert1, exponent)
}
\arguments{
\item{pauli0}{The interaction axis for the first qubit.}

\item{invert0}{bool. Whether to condition on the +1 or -1
eigenvector of the first qubit’s interaction axis.}

\item{pauli1}{The interaction axis for the second qubit.}

\item{invert1}{bool. Whether to condition on the +1 or -1
eigenvector of the second qubit’s interaction axis.}

\item{exponent}{float. Determines the amount of phasing to
apply to the vector equal to the tensor product of the two conditions.}
}
\description{
A CZ conjugated by arbitrary single qubit Cliffords.
}
\seealso{
Other Pauli and Clifford groups: 
\code{\link{clifford_state}()},
\code{\link{clifford_tableau}()},
\code{\link{pauli_base_dense_string}()},
\code{\link{pauli_basis}()},
\code{\link{pauli_dense_string}()},
\code{\link{pauli_gate_like}()},
\code{\link{pauli_mutable_dense_string}()},
\code{\link{pauli_pow_combination}()},
\code{\link{pauli}()}
}
\concept{Pauli and Clifford groups}
