% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pauli_Clifford.R
\name{pauli_string}
\alias{pauli_string}
\title{Pauli String}
\usage{
pauli_string(...)
}
\arguments{
\item{...}{parameters to pass:
- `*contents` A value or values to convert into a pauli string.
This can be a number, a pauli operation, a dictionary from qubit
to pauli/identity gates, or collections thereof. If a list of values
is given, they are each individually converted and then multiplied
from left to right in order.
- qubit_pauli_map Initial dictionary mapping qubits to pauli
operations. Defaults to the empty dictionary. Note that, unlike
dictionaries passed to contents, this dictionary must not contain any
identity gate values. Further note that this argument specifies values
that are logically before factors specified in contents; contents are
right multiplied onto the values in this dictionary.
- coefficient Initial scalar coefficient. Defaults to 1.}
}
\value{
a pauli string object.
}
\description{
Initializes a new PauliString.
}
\seealso{
Other Pauli and Clifford groups: 
\code{\link{clifford_single_qubit_gate}()},
\code{\link{clifford_state}()},
\code{\link{clifford_tableau}()},
\code{\link{pauli_base_dense_string}()},
\code{\link{pauli_basis}()},
\code{\link{pauli_dense_string}()},
\code{\link{pauli_gate_like}()},
\code{\link{pauli_interaction_gate}()},
\code{\link{pauli_mutable_dense_string}()},
\code{\link{pauli_pow_combination}()},
\code{\link{pauli_string_gate_operation}()},
\code{\link{pauli_sum_like}()},
\code{\link{pauli_sum}()},
\code{\link{pauli_transform}()},
\code{\link{pauli}()}
}
\concept{Pauli and Clifford groups}
