% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pauli_Clifford.R
\name{pauli_sum}
\alias{pauli_sum}
\title{Pauli Sum}
\usage{
pauli_sum(linear_dict = NULL)
}
\arguments{
\item{linear_dict}{Under the hood, this class is backed by a LinearDict with coefficient-less
PauliStrings as keys. PauliStrings are reconstructed on-the-fly during
iteration. PauliStrings are reconstructed on-the-fly during
iteration.}
}
\value{
an operation with pauli sum
}
\description{
Represents operator defined by linear combination of PauliStrings.
}
\details{
Since PauliStrings store their own coefficients, this class
does not implement the LinearDict interface. Instead, you can
add and subtract terms and then iterate over the resulting
(simplified) expression.
}
\seealso{
Other Pauli and Clifford groups: 
\code{\link{clifford_single_qubit_gate}()},
\code{\link{clifford_state}()},
\code{\link{clifford_tableau}()},
\code{\link{pauli_base_dense_string}()},
\code{\link{pauli_basis}()},
\code{\link{pauli_dense_string}()},
\code{\link{pauli_gate_like}()},
\code{\link{pauli_interaction_gate}()},
\code{\link{pauli_mutable_dense_string}()},
\code{\link{pauli_pow_combination}()},
\code{\link{pauli_string_gate_operation}()},
\code{\link{pauli_string}()},
\code{\link{pauli_sum_like}()},
\code{\link{pauli_transform}()},
\code{\link{pauli}()}
}
\concept{Pauli and Clifford groups}
