% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pauli_Clifford.R
\name{pauli_transform}
\alias{pauli_transform}
\title{Pauli Transform}
\usage{
pauli_transform(...)
}
\arguments{
\item{...}{parameters to pass.}
}
\value{
sum of pauli arguments.
}
\description{
Any value that can be easily transofrmed into a sum of Pauli products.
+X, -X, +Y, -Y, +Z, or -Z.
}
\seealso{
Other Pauli and Clifford groups: 
\code{\link{clifford_single_qubit_gate}()},
\code{\link{clifford_state}()},
\code{\link{clifford_tableau}()},
\code{\link{pauli_base_dense_string}()},
\code{\link{pauli_basis}()},
\code{\link{pauli_dense_string}()},
\code{\link{pauli_gate_like}()},
\code{\link{pauli_interaction_gate}()},
\code{\link{pauli_mutable_dense_string}()},
\code{\link{pauli_pow_combination}()},
\code{\link{pauli_string_gate_operation}()},
\code{\link{pauli_string}()},
\code{\link{pauli_sum_like}()},
\code{\link{pauli_sum}()},
\code{\link{pauli}()}
}
\concept{Pauli and Clifford groups}
