% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_print_op.R
\name{print.cirq.circuits.circuit.Circuit}
\alias{print.cirq.circuits.circuit.Circuit}
\title{Circuit}
\usage{
\method{print}{cirq.circuits.circuit.Circuit}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Circuit
}
