% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.cirq.devices.line_qubit.LineQubit}
\alias{print.cirq.devices.line_qubit.LineQubit}
\title{LineQubit}
\usage{
\method{print}{cirq.devices.line_qubit.LineQubit}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
LineQubit
}
