% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_print_op.R
\name{print.cirq.ops.common_gates.XPowGate}
\alias{print.cirq.ops.common_gates.XPowGate}
\title{XPowGate}
\usage{
\method{print}{cirq.ops.common_gates.XPowGate}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
XPowGate
}
