% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.cirq.ops.pauli_gates._PauliZ}
\alias{print.cirq.ops.pauli_gates._PauliZ}
\title{_PauliZ}
\usage{
\method{print}{cirq.ops.pauli_gates._PauliZ}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
_PauliZ
}
