% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_print_op.R
\name{print.cirq.ops.pauli_string.PauliString}
\alias{print.cirq.ops.pauli_string.PauliString}
\title{PauliString}
\usage{
\method{print}{cirq.ops.pauli_string.PauliString}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
PauliString
}
