% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R
\name{print.cirq.ops.swap_gates.ISwapPowGate}
\alias{print.cirq.ops.swap_gates.ISwapPowGate}
\title{ISwapPowGate}
\usage{
\method{print}{cirq.ops.swap_gates.ISwapPowGate}(x, ...)
}
\arguments{
\item{x}{an object used to select a method.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
ISwapPowGate
}
