% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Devices.R
\name{qid}
\alias{qid}
\title{Qid}
\usage{
qid(...)
}
\arguments{
\item{...}{parameters to pass.}
}
\value{
qid object.
}
\description{
Identifies a quantum object such as a qubit, qudit, resonator, etc.
}
\details{
Child classes represent specific types of objects, such as a qubit at a
particular location on a chip or a qubit with a particular name.
The main criteria that a custom qid must satisfy is comparability. Child
classes meet this criteria by implementing the _comparison_key method. For
example, cirq.LineQubit’s _comparison_key method returns self.x. This
ensures that line qubits with the same x are equal, and that line qubits
will be sorted ascending by x. Qid implements all equality,
comparison, and hashing methods via _comparison_key.
}
\seealso{
Other Devices: 
\code{\link{grid_qid}()},
\code{\link{grid_qubit}()},
\code{\link{line_qid}()},
\code{\link{line_qubit}()},
\code{\link{named_qubit}()},
\code{\link{unconstrained_device}()}
}
\concept{Devices}
