% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qubits.R
\name{qubit_grid}
\alias{qubit_grid}
\title{Grid Qubit}
\usage{
qubit_grid(row, col)
}
\arguments{
\item{row}{row number of a rectangular lattice.}

\item{col}{column number of a rectangular lattice.}
}
\description{
A qubit on a 2d square lattice.
}
\details{
GridQubits use row-major ordering:
GridQubit(0, 0) < GridQubit(0, 1) < GridQubit(1, 0) < GridQubit(1, 1)
New GridQubits can be constructed by adding or subtracting lists
>>> cirq.GridQubit(2, 3) + (3, 1) cirq.GridQubit(5, 4) >>> cirq.GridQubit(2, 3) - (1, 2) cirq.GridQubit(1, 1)
}
