% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qubits.R
\name{qubit_grid_square}
\alias{qubit_grid_square}
\title{Grid Qubits square}
\usage{
qubit_grid_square(diameter, top = 0, left = 0)
}
\arguments{
\item{diameter}{Length of a side of the square}

\item{top}{Row number of the topmost row}

\item{left}{Column number of the leftmost row}
}
\value{
A list of GridQubits filling in a square grid
}
\description{
Returns a square of GridQubits.
}
