% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qubits.R
\name{qubit_line_range}
\alias{qubit_line_range}
\title{Line Qubit range}
\usage{
qubit_line_range(...)
}
\arguments{
\item{...}{the length of parameter. E.g. the same with R base
function `seq(1,5,1)`. It returns the range in the form of list.}
}
\value{
a range of line qubits
}
\description{
Line Qubit range
}
