% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qubits.R
\name{qubit_named}
\alias{qubit_named}
\title{Named Qubit}
\usage{
qubit_named(name)
}
\arguments{
\item{name}{the name of the op}
}
\description{
A qubit identified by name.
}
\details{
By default, NamedQubit has a lexicographic order. However, numbers within
the name are handled correctly. So, for example, if you print a circuit
containing `qubit_named('qubit22')` and `qubit_named('qubit3')`, the
wire for 'qubit3' will correctly come before 'qubit22'.
}
