% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circuits.R
\name{qubit_order_or_list}
\alias{qubit_order_or_list}
\title{Qubit Order or List}
\usage{
qubit_order_or_list(...)
}
\arguments{
\item{...}{parameters to pass.}
}
\value{
None
}
\description{
Defines the kronecker product order of qubits.
}
\seealso{
Other Circuits: 
\code{\link{circuit_dag}()},
\code{\link{circuit}()},
\code{\link{gate_operation_}()},
\code{\link{gate_parallel_operation}()},
\code{\link{insert_strategy}()},
\code{\link{moment}()},
\code{\link{qubit_order}()},
\code{\link{tree_flatten_op}()},
\code{\link{tree_op}()}
}
\concept{Circuits}
