% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Importing_Exporting.R
\name{quirk_json_to_circuit}
\alias{quirk_json_to_circuit}
\title{Quirk json to circuit}
\usage{
quirk_json_to_circuit(
  data,
  qubits,
  extra_cell_makers,
  quirk_url,
  max_operation_count
)
}
\arguments{
\item{data}{Data parsed from quirk's JSON representation.}

\item{qubits}{Qubits to use in the circuit. See quirk_url_to_circuit.}

\item{extra_cell_makers}{Non-standard Quirk cells to accept.
See quirk_url_to_circuit.}

\item{quirk_url}{If given, the original URL from which the
JSON was parsed, as described in quirk_url_to_circuit.}

\item{max_operation_count}{If the number of operations
in the circuit would exceed this value, the method raises
a ValueError instead of attempting to construct the circuit.
This is important to specify for servers parsing unknown input,
because Quirk’s format allows for a billion laughs attack in
the form of nested custom gates.}
}
\value{
The parsed circuit.
}
\description{
Constructs a Cirq circuit from Quirk's JSON format.
}
\section{Raises}{

ValueError: Invalid circuit URL, or circuit would be larger than `max_operations_count`.
}

\seealso{
Other Importing and Exporting: 
\code{\link{quirk_url_to_circuit}()}
}
\concept{Importing and Exporting}
