% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_simulations.R
\name{random_state_or_seed_like}
\alias{random_state_or_seed_like}
\title{Random state or seed like}
\usage{
random_state_or_seed_like(...)
}
\arguments{
\item{...}{parameters to pass}
}
\value{
None
}
\description{
A pseudorandom number generator or object that can
be converted to one.
}
\details{
If NULL, turns into the module np.random.

If an integer, turns into a np.random.RandomState seeded with that
integer.
If none of the above, it is used unmodified. In this case, it is assumed
that the object implements whatever methods are required for the use case
at hand. For example, it might be an existing instance of
np.random.RandomState or a custom pseudorandom number generator
implementation.
}
\seealso{
Other Sampling, Simulations and Data collection: 
\code{\link{big_endian_bits_to_int}()},
\code{\link{big_endian_digits_to_int}()},
\code{\link{big_endian_int_to_bits}()},
\code{\link{big_endian_int_to_digits}()},
\code{\link{circuit_like}()},
\code{\link{final_density_matrix}()},
\code{\link{final_wavefunction}()},
\code{\link{flatten_to_ops_or_moments}()},
\code{\link{flatten_to_ops}()},
\code{\link{flatten_with_params}()},
\code{\link{flatten_with_sweep}()},
\code{\link{flatten}()},
\code{\link{hog_score_xeb_fidelity_from_probabilities}()},
\code{\link{measure_density_matrix}()},
\code{\link{measure_state_vector}()},
\code{\link{sample_density_matrix}()},
\code{\link{sample_state_vector}()},
\code{\link{sample_sweep}()},
\code{\link{sample}()},
\code{\link{to_resolvers}()},
\code{\link{to_sweep}()},
\code{\link{validate_mixture}()}
}
\concept{Sampling, Simulations and Data collection}
