% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_simulations.R
\name{sample}
\alias{sample}
\title{Sample}
\usage{
sample(program, ...)
}
\arguments{
\item{program}{The circuit to sample from.}

\item{...}{additional parameters to pass.}
}
\description{
Simulates sampling from the given circuit.
}
\seealso{
Other Sampling, Simulations and Data collection: 
\code{\link{big_endian_bits_to_int}()},
\code{\link{big_endian_digits_to_int}()},
\code{\link{big_endian_int_to_bits}()},
\code{\link{big_endian_int_to_digits}()},
\code{\link{circuit_like}()},
\code{\link{final_density_matrix}()},
\code{\link{final_wavefunction}()},
\code{\link{flatten_to_ops_or_moments}()},
\code{\link{flatten_to_ops}()},
\code{\link{flatten_with_params}()},
\code{\link{flatten_with_sweep}()},
\code{\link{flatten}()},
\code{\link{hog_score_xeb_fidelity_from_probabilities}()},
\code{\link{measure_density_matrix}()},
\code{\link{measure_state_vector}()},
\code{\link{random_state_or_seed_like}()},
\code{\link{sample_density_matrix}()},
\code{\link{sample_state_vector}()},
\code{\link{sample_sweep}()},
\code{\link{to_resolvers}()}
}
\concept{Sampling, Simulations and Data collection}
