% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_simulations.R
\name{sample_state_vector}
\alias{sample_state_vector}
\title{Sample state vector}
\usage{
sample_state_vector(state, indices)
}
\arguments{
\item{state}{The multi-qubit wavefunction to be sampled. This is an array of
2 to the power of the number of qubit complex numbers, and so state must be of
size ``2**integer``. The state can be a vector of size ``2**integer`` or a
tensor of shape ``(2, 2, ..., 2)``.}

\item{indices}{Which qubits are measured. The state is assumed to be supplied
in big endian order. That is the xth index of v, when expressed as a bitstring,
has its largest values in the 0th index.}
}
\value{
Measurement results with TRUE corresponding to the ``|1⟩`` state. The outer
list is for repetitions, and the inner corresponds to measurements ordered by the
supplied qubits. These lists are wrapped as an numpy ndarray.
}
\description{
Samples repeatedly from measurements in the computational basis.
}
\details{
Note that this does not modify the passed in state.
}
\section{Raises}{

ValueError: ``repetitions`` is less than one or size of ``state`` is not
a power of 2. IndexError: An index from ``indices`` is out of range, given the
number of qubits corresponding to the state.
}

\seealso{
Other Sampling, Simulations and Data collection: 
\code{\link{big_endian_bits_to_int}()},
\code{\link{big_endian_digits_to_int}()},
\code{\link{big_endian_int_to_bits}()},
\code{\link{big_endian_int_to_digits}()},
\code{\link{circuit_like}()},
\code{\link{final_density_matrix}()},
\code{\link{final_wavefunction}()},
\code{\link{flatten_to_ops_or_moments}()},
\code{\link{flatten_to_ops}()},
\code{\link{flatten_with_params}()},
\code{\link{flatten_with_sweep}()},
\code{\link{flatten}()},
\code{\link{hog_score_xeb_fidelity_from_probabilities}()},
\code{\link{measure_density_matrix}()},
\code{\link{measure_state_vector}()},
\code{\link{random_state_or_seed_like}()},
\code{\link{sample_density_matrix}()},
\code{\link{sample_sweep}()},
\code{\link{sample}()},
\code{\link{to_resolvers}()}
}
\concept{Sampling, Simulations and Data collection}
