% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pauli_Clifford.R
\name{stabilizer_state_ch_form}
\alias{stabilizer_state_ch_form}
\title{StabilizerStateChForm}
\usage{
stabilizer_state_ch_form(num_qubits, initial_state = 0, ...)
}
\arguments{
\item{num_qubits}{The number of qubits in the system}

\item{initial_state}{If an int, the state is set to the computational}

\item{...}{other arguments to pass.}
}
\value{
initialized stabilizer states using the CH form.
}
\description{
A representation of stabilizer states using the CH form.
}
\details{
$|\\psi> = \\omega U_C U_H |s>$ This representation keeps track of overall phase.
Reference: https://arxiv.org/abs/1808.00128
}
