% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_print_op.R
\name{str.cirq.circuits.circuit.Circuit}
\alias{str.cirq.circuits.circuit.Circuit}
\title{Circuit}
\usage{
\method{str}{cirq.circuits.circuit.Circuit}(object, ...)
}
\arguments{
\item{object}{any R object about which you want to have some information.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Circuit
}
