% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circuits.R
\name{tree_flatten_op}
\alias{tree_flatten_op}
\title{Flatten op tree}
\usage{
tree_flatten_op(root, preserve_moments = FALSE)
}
\arguments{
\item{root}{The operation or tree of operations to iterate.}

\item{preserve_moments}{Whether to yield Moments intact instead of flattening them.}
}
\value{
Operations from the tree.
}
\description{
Performs an in-order iteration of the operations (leaves) in an OP_TREE.
}
\section{Raises}{

TypeError: root isn't a valid OP_TREE.
}

\seealso{
Other Circuits: 
\code{\link{circuit_dag}()},
\code{\link{circuit}()},
\code{\link{gate_operation_}()},
\code{\link{gate_parallel_operation}()},
\code{\link{insert_strategy}()},
\code{\link{moment}()},
\code{\link{qubit_order_or_list}()},
\code{\link{qubit_order}()},
\code{\link{tree_op}()}
}
\concept{Circuits}
