% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circuits.R
\name{tree_freeze_op}
\alias{tree_freeze_op}
\title{Freeze op tree}
\usage{
tree_freeze_op(root)
}
\arguments{
\item{root}{The operation or tree of operations to freeze.}
}
\value{
An OP_TREE with the same operations and branching structure,
but where all internal nodes are lists instead of arbitrary iterables.

An OP_TREE with the same operations and branching structure, but
where all internal nodes are list of multiple types instead of
arbitrary iterables.
}
\description{
Replaces all iterables in the OP_TREE with lists.
}
