% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circuits.R
\name{tree_op}
\alias{tree_op}
\title{Op tree}
\usage{
tree_op(...)
}
\arguments{
\item{...}{parameters to pass.}
}
\value{
operations
}
\description{
An operation or nested collections of operations.
}
\seealso{
Other Circuits: 
\code{\link{circuit_dag}()},
\code{\link{circuit}()},
\code{\link{gate_operation_}()},
\code{\link{gate_parallel_operation}()},
\code{\link{insert_strategy}()},
\code{\link{moment}()},
\code{\link{qubit_order_or_list}()},
\code{\link{qubit_order}()},
\code{\link{tree_flatten_op}()}
}
\concept{Circuits}
