% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Circuits.R
\name{tree_transform_op}
\alias{tree_transform_op}
\title{Transform op tree}
\usage{
tree_transform_op(
  root,
  op_transformation,
  iter_transformation,
  preserve_moments = FALSE
)
}
\arguments{
\item{root}{The operation or tree of operations to transform.}

\item{op_transformation}{How to transform the operations
(i.e. leaves).}

\item{iter_transformation}{How to transform the iterables
(i.e. internal nodes).}

\item{preserve_moments}{Whether to leave Moments alone. If TRUE,
the transformation functions will not be applied to Moments or
the operations within them.}
}
\value{
A transformed operation tree.
}
\description{
Maps transformation functions onto the nodes of an OP_TREE.
}
\section{Raises}{

TypeError: root isn't a valid OP_TREE.
}

