% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Devices.R
\name{unconstrained_device}
\alias{unconstrained_device}
\title{Unconstrained device}
\usage{
unconstrained_device(...)

unconstrained_device(...)
}
\arguments{
\item{...}{parameters to pass.}
}
\value{
None

device with options
}
\description{
A device with no constraints on operations or qubits.

Hardware constraints for validating circuits.
}
\seealso{
Other Devices: 
\code{\link{grid_qid}()},
\code{\link{grid_qubit}()},
\code{\link{line_qid}()},
\code{\link{line_qubit}()},
\code{\link{named_qubit}()},
\code{\link{qid}()}

Other Devices: 
\code{\link{grid_qid}()},
\code{\link{grid_qubit}()},
\code{\link{line_qid}()},
\code{\link{line_qubit}()},
\code{\link{named_qubit}()},
\code{\link{qid}()}
}
\concept{Devices}
