% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_simulations.R
\name{validate_probability}
\alias{validate_probability}
\title{Validate probability}
\usage{
validate_probability(p, p_str)
}
\arguments{
\item{p}{The value to validate.}

\item{p_str}{What to call the probability in error messages.}
}
\value{
The probability p if the probability if valid.
}
\description{
Validates that a probability is between 0 and 1 inclusively.
}
\section{Raises}{

ValueError if the probability is invalid.
}

