% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling_simulations.R
\name{xeb_fidelity}
\alias{xeb_fidelity}
\title{Xeb fidelity}
\usage{
xeb_fidelity(circuit, bitstrings, qubit_order, amplitudes = NULL, estimator)
}
\arguments{
\item{circuit}{Random quantum circuit which has been executed on quantum processor under test.}

\item{bitstrings}{Results of terminal all-qubit measurements performed after
each circuit execution as integer array where each integer is formed from
measured qubit values according to `qubit_order` from most to least significant
qubit, i.e. in the order consistent with `cirq.final_wavefunction`.}

\item{qubit_order}{Qubit order used to construct bitstrings enumerating qubits
starting with the most sigificant qubit.}

\item{amplitudes}{Optional mapping from bitstring to output amplitude. If provided,
simulation is skipped. Useful for large circuits when an offline simulation
had already been peformed.}

\item{estimator}{Fidelity estimator to use, see above. Defaults to the linear
XEB fidelity estimator. Returns: Estimate of fidelity associated with an experimental
realization of circuit which yielded measurements in bitstrings. Raises:}
}
\value{
Estimate of fidelity associated with an experimental realization of
circuit which yielded measurements in bitstrings. Raises: ValueError: Circuit
is inconsistent with qubit order or one of the bitstrings is inconsistent with the number of qubits.
}
\description{
Estimates XEB fidelity from one circuit using user-supplied estimator.
}
\details{
Fidelity quantifies the similarity of two quantum states. Here, we estimate
the fidelity between the theoretically predicted output state of circuit and
the state producted in its experimental realization. Note that we don't know
the latter state. Nevertheless, we can estimate the fidelity between the two
states from the knowledge of the bitstrings observed in the experiment.
In order to make the estimate more robust one should average the estimates
over many random circuits. The API supports per-circuit fidelity estimation
to enable users to examine the properties of estimate distribution over
many circuits. See https://arxiv.org/abs/1608.00263 for more details.
}
\section{Raises}{

ValueError: Circuit is inconsistent with
qubit order or one of the bitstrings is inconsistent with the number of qubits.
}

